refactor: 필터링된 음식점 목록에 대한 로직 Store에서 분리
- Store 에는 음식점 목록과, 필터링 옵션 값들, 선택된 음식점에 대한 정보만 관리하도록 수정
- 필터링된 음식점 목록은 Store에서 관리하는 것보다 필터링된 음식점 목록이 필요한 부분에서 사용하여 UI 업데이트 해주도록 함
- getFilteredRestaurants 메서드에서 필터링 옵션에 대한 기본값이 설정되도록 함
- 처음 필터링된 음식점 목록을 UI에 렌더링할 때는 가장 기본값인 '모든 음식점 보기', '전체', '이름순'으로 설정
